+++ src/components/profile-navigation/profile-navigation.module.css
@@ -1,0 +1,30 @@
+
+.navigationWrapper {
+    display: flex;
+    flex-direction: column;
+    width: 25%;
+}
+
+.profileLink {
+    height: 60px;
+    display: flex;
+    align-items: center;
+}
+
+    .profileLink a {
+        text-decoration: none;
+        color: #8585AD;
+    }
+
+.profileLinkActive {
+    color: white !important;
+}
+
+
+
+.logoutBtn {
+    border: none;
+    background-color: transparent;
+    cursor: pointer;
+    color: #8585AD;
+}
