+++ src/components/profile-orders/profile-orders.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './profile-orders.module.css';
+
+export default function OrdersHistory() {
+    return (
+        <div>
+            TODO
+        </div>
+        );
+}
