+++ src/components/profile/profile.module.css
@@ -1,0 +1,33 @@
+.input > div > div {
+    width: 380px;
+}
+
+
+.contentWrapper {
+    display: flex;
+    flex-direction: column;
+    width: 75%;
+}
+
+
+.errorMessage {
+    color: red;
+}
+
+.buttonsWrapper {
+    display: flex;
+    flex-direction: row;
+    justify-content: right;
+    width: 380px;
+    align-items: center;
+}
+
+.cancelBtn {
+    padding: 0;
+    border: none;
+    font: inherit;
+    color: inherit;
+    background-color: transparent;
+    color: #4C4CFF;
+    cursor: pointer;
+}
