+++ src/pages/profile.module.css
@@ -3,67 +3,4 @@
     flex-direction: row;
 }
 
-.navigationWrapper {
-    display: flex;
-    flex-direction: column;
-    width: 25%;
-}
 
-.rightWrapper {
-    display: flex;
-    flex-direction: column;
-}
-
-.contentWrapper {
-    display: flex;
-    flex-direction: column;
-    width: 75%;
-}
-
-.profileLink {
-    height: 60px;
-    display: flex;
-    align-items: center;
-}
-
-    .profileLink a {
-        text-decoration: none;
-        color: #8585AD;
-    }
-
-.profileLinkActive {
-    color: white !important;
-}
-
-.input>div>div {
-    width: 380px;
-}
-
-.errorMessage {
-    color: red;
-}
-
-.buttonsWrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: right;
-    width: 380px;
-    align-items: center;
-}
-
-.cancelBtn {
-    padding: 0;
-    border: none;
-    font: inherit;
-    color: inherit;
-    background-color: transparent;
-    color: #4C4CFF;
-    cursor: pointer;
-}
-
-.logoutBtn {
-    border: none;
-    background-color: transparent;
-    cursor: pointer;
-    color: #8585AD;
-}
